Revert "Bug fix"
This reverts commit 8b136ce6229b5f592e1b385898e5f57571389ec7.

+++ src/br/com/mvbos/mymer/sync/Differ.java
@@ -234,7 +234,7 @@ public class Differ {
         entityToScript.setMode(IEntityToScript.Mode.PLAIN);
         entityToScript.addTable(te, sb);
 
-        short ct = 1;
+        int ct = 1;
 
         for (Field f : te.getFields()) {
             entityToScript.addField(te, f, ct++, sb);
